+++ EpLibrary2.0/EpLibrary/Headers/epThreadSafePQueue.h
@@ -114,7 +114,7 @@ namespace epl
 		DataType* retVal=NULL;
 		if(m_queue.size())
 		{
-			int retIdx;
+			size_t retIdx;
 			if(BinarySearch(data,&m_queue.at(0),m_queue.size(),Compare::CompFunc,retIdx))
 				EP_ASSERT_EXPR(0,_T("Same Object already in the Queue!!"));
 			m_queue.insert(m_queue.begin()+retIdx,data);
